improvement: separated users from requests, reverting userref->null
