+++ functions/src/https/api/index.ts
@@ -1,4 +1,5 @@
 import * as map from './map';
 import * as requests from './requests';
+import * as users from './users';
 
-export { map, requests };
+export { map, requests, users };
